fix: BigInteger for SNOMED codes and Metriport consolidated query
- Change omop_concept_id from Integer to BigInteger in 4 models
  (clinical_fact, knowledge_graph, mention, clinical_value) to handle
  SNOMED codes that exceed int32 range (e.g. 428211000124100)
- Fix consolidated query endpoint from GET to POST with patient ID in
  path instead of query params
- Fix seed script connectivity check to use get_facility instead of
  get_organization (unsupported in sandbox)
- Add sandbox patient results JSON from successful E2E test

+++ backend/app/models/clinical_fact.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import DateTime, Enum, Float, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, Enum, Float, ForeignKey, Index, Integer, String, Text
 from sqlalchemy.dialects.postgresql import ARRAY, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -42,7 +42,7 @@ class ClinicalFact(SoftDeleteMixin, Base):
         index=True,
     )
     omop_concept_id: Mapped[int] = mapped_column(
-        Integer,
+        BigInteger,
         nullable=False,
         index=True,
     )
